🏗️ migrate necessary files to build through vite
install vite dependencies, change environment keys,
move all .js components to .jsx, fix internal imports,
fix modules imports and change package.json scripts section
