+++ .env.development.example
@@ -1,6 +1,6 @@
 # WARNING: Develoment server does not autoreload this file, requires a restart
 
-REACT_APP_DEVEL_TOKEN=super_secret_token
-REACT_APP_DEVEL_CONTRACTS='[{"name": "0012345", "cups": "ES0031406666910014WM0F", "tariff": "2.0TD", "address": "13 Rue del Percebe", "tariff_timetable_id": "Taula_Peatges_20"}, {"name": "0054321", "cups": "ES0031666666566001YH0F", "tariff": "2.0TD", "address": "Palacio de la Zarzuela s/n (Madrid)", "tariff_timetable_id": "Taula_Peatges_20"}]'
-REACT_APP_HEMAN_API_URL=httx://heman_api_url.coop
-REACT_APP_WEBFORMS_API_URL=httx://api_url.coop
+VITE_APP_DEVEL_TOKEN=super_secret_token
+VITE_APP_DEVEL_CONTRACTS='[{"name": "0012345", "cups": "ES0031406666910014WM0F", "tariff": "2.0TD", "address": "13 Rue del Percebe", "tariff_timetable_id": "Taula_Peatges_20"}, {"name": "0054321", "cups": "ES0031666666566001YH0F", "tariff": "2.0TD", "address": "Palacio de la Zarzuela s/n (Madrid)", "tariff_timetable_id": "Taula_Peatges_20"}]'
+VITE_APP_HEMAN_API_URL=httx://heman_api_url.coop
+VITE_APP_WEBFORMS_API_URL=httx://api_url.coop
