+++ .gitignore
@@ -22,6 +22,7 @@ __pycache__
 
 # misc
 .DS_Store
+.env
 .env.local
 .env.development.local
 .env.development
@@ -32,3 +33,6 @@ npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
 public/infoenergia-config.js
+
+# others
+bun*
