+++ CHANGES.md
@@ -7,10 +7,10 @@
 ## 1.3.0 2025-01-22
 - Improve investments Api, divided in two parts, mokc and production
 - Add new env variables:
-    - REACT_APP_GENERATION_ASSIGNMENTS_USE_TABLE
-    - REACT_APP_GENERATION_PRODUCTION_CHART
-    - REACT_APP_GENERATION_BAG_KWH
-    - REACT_APP_GENERATION_LAST_INVOICE_DATE 
+    - VITE_APP_GENERATION_ASSIGNMENTS_USE_TABLE
+    - VITE_APP_GENERATION_PRODUCTION_CHART
+    - VITE_APP_GENERATION_BAG_KWH
+    - VITE_APP_GENERATION_LAST_INVOICE_DATE 
 
 ## 1.2.0 2024-08-20
