+++ index.html
@@ -2,19 +2,19 @@
 <html lang="en">
   <head>
     <meta charset="utf-8" />
-    <link rel="icon" href="%PUBLIC_URL%/favicon.ico" />
+    <link rel="icon" href="/favicon.ico" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <meta name="theme-color" content="#000000" />
     <meta
       name="description"
       content="Web site created using create-react-app"
     />
-    <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" />
+    <link rel="apple-touch-icon" href="/logo192.png" />
     <!--
       manifest.json provides metadata used when your web app is installed on a
       user's mobile device or desktop. See https://developers.google.com/web/fundamentals/web-app-manifest/
     -->
-    <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
+    <link rel="manifest" href="/manifest.json" />
     <link rel="preconnect" href="https://fonts.gstatic.com" />
     <link
       href="https://fonts.googleapis.com/css2?family=Roboto:wght@400;500;700;800&display=swap"
@@ -22,17 +22,17 @@
     />
 
     <!--
-      Notice the use of %PUBLIC_URL% in the tags above.
+      Notice the use of  in the tags above.
       It will be replaced with the URL of the `public` folder during the build.
       Only files inside the `public` folder can be referenced from the HTML.
 
-      Unlike "/favicon.ico" or "favicon.ico", "%PUBLIC_URL%/favicon.ico" will
+      Unlike "/favicon.ico" or "favicon.ico", "/favicon.ico" will
       work correctly both with client-side routing and a non-root public URL.
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
 
     <!-- runtime config -->
-    <script src="%PUBLIC_URL%/infoenergia-config.js"></script>
+    <script src="/infoenergia-config.js"></script>
     <!-- runtime config -->
     <title>InfoEnergia Dashboard</title>
   </head>
@@ -41,26 +41,17 @@
     <!-- dev token: 04afcfa463934dc784540aaea63d7d76490101cb0e264c43a7c293bdf58a17d4 -->
     <!-- pro token: ecada804a8884a67960726dbf983aeda451355d1ae2d4af69854dd7e0d182761 -->
     <script id="contracts-data" type="application/json">
-      %REACT_APP_DEVEL_CONTRACTS%
+      %VITE_APP_DEVEL_CONTRACTS%
     </script>
     <script id="generation-assignments-data" type="application/json">
-      %REACT_APP_GENERATION_ASSIGNMENTS%
+      %VITE_APP_GENERATION_ASSIGNMENTS%
     </script>
     <div
       id="root"
-      data-token="%REACT_APP_DEVEL_TOKEN%"
-      data-heman-api-url="%REACT_APP_HEMAN_API_URL%"
-      data-webforms-api-url="%REACT_APP_WEBFORMS_API_URL%"
+      data-token="%VITE_APP_DEVEL_TOKEN%"
+      data-heman-api-url="%VITE_APP_HEMAN_API_URL%"
+      data-webforms-api-url="%VITE_APP_WEBFORMS_API_URL%"
     ></div>
-    <!--
-      This HTML file is a template.
-      If you open it directly in the browser, you will see an empty page.
-
-      You can add webfonts, meta tags, or analytics to this file.
-      The build step will place the bundled scripts into the <body> tag.
-
-      To begin the development, run `npm start` or `yarn start`.
-      To create a production bundle, use `npm run build` or `yarn build`.
-    -->
+    <script type="module" src="/src/main.jsx"></script>
   </body>
 </html>
