+++ infoenergia-config.js.example
@@ -1,0 +1,2 @@
+window.config = {
+}
