+++ package.json
@@ -12,6 +12,7 @@
     "@testing-library/jest-dom": "^5.16.2",
     "@testing-library/react": "^12.1.4",
     "@testing-library/user-event": "^13.5.0",
+    "@vitejs/plugin-react": "^5.1.1",
     "axios": "^1.6.5",
     "date-fns": "^2.30.0",
     "dayjs": "^1.11.0",
@@ -23,11 +24,12 @@
     "react-i18next": "^11.15.7",
     "react-router": "^5.1.2",
     "react-router-dom": "^5.1.2",
-    "react-scripts": "^5.0.0",
     "recharts": "^2.1.9",
-    "styled-components": "^5.3.3"
+    "styled-components": "^5.3.3",
+    "vite": "^7.2.6"
   },
   "devDependencies": {
+    "@types/react-router-dom": "^5.3.3",
     "env-cmd": "^10.1.0",
     "eslint": "^7.22.0",
     "eslint-config-prettier": "^8.1.0",
@@ -48,11 +50,11 @@
     ]
   },
   "scripts": {
-    "start": "export REACT_APP_VERSION=$(git rev-parse HEAD) && react-scripts start",
-    "build": "export REACT_APP_VERSION=$(git rev-parse HEAD) && react-scripts build",
-    "build:ov": "env-cmd -f .env.ov  --use-shell \"npm run build && ./scripts/rename-assets-ov.sh\"",
+    "dev": "export VITE_APP_VERSION=$(git rev-parse HEAD) && vite",
+    "build": "export VITE_APP_VERSION=$(git rev-parse HEAD) && vite build",
     "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "eject": "react-scripts eject",
+    "build:ov": "env-cmd -f .env.ov  --use-shell \"npm run build && ./scripts/rename-assets-ov.sh\""
   },
   "jest": {
     "moduleNameMapper": {
