+++ src/components/Generation/PeriodSelector.jsx
