+++ src/components/Generation/StackedBarChart.jsx
