+++ src/components/Loading.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { makeStyles } from '@material-ui/core/styles'
 
-import loading from 'images/loading.svg'
+import loading from '../images/loading.svg'
 
 const useStyles = makeStyles((theme) => ({
   loading: {
