+++ src/components/SeasonalProfile/ClimaDependency.jsx
