+++ src/components/SeasonalProfile/SeasonalProfileBarChart.jsx
