+++ src/components/TipicalDailyProfile/DistributionLegend.jsx
