+++ src/components/TipicalDailyProfile/DistributionPieChart.jsx
