+++ src/components/TipicalDailyProfile/LegendPeriod.jsx
