+++ src/components/TipicalDailyProfile/LegendTable.jsx
