+++ src/components/TipicalDailyProfile/TipicalDailyProfileChart.jsx
