+++ src/components/TipicalWeeklyProfile/TipicalWeeklyProfileChart.jsx
