+++ src/containers/TipicalDailyProfile.jsx
@@ -7,17 +7,17 @@ import Skeleton from '@material-ui/lab/Skeleton'
 import Typography from '@material-ui/core/Typography'
 import { makeStyles } from '@material-ui/core/styles'
 
-import TipicalDailyProfileChart from 'components/TipicalDailyProfile/TipicalDailyProfileChart'
-import Counter from 'components/Counter'
-import LastUpdate from 'components/LastUpdate'
+import TipicalDailyProfileChart from '../components/TipicalDailyProfile/TipicalDailyProfileChart'
+import Counter from '../components/Counter'
+import LastUpdate from '../components/LastUpdate'
 
-import DistributionByPeriod from 'containers/TipicalDailyProfile/DistributionByPeriod'
-import DistributionByUserType from 'containers/TipicalDailyProfile/DistributionByUserType'
+import DistributionByPeriod from './TipicalDailyProfile/DistributionByPeriod'
+import DistributionByUserType from './TipicalDailyProfile/DistributionByUserType'
 
-import { ScrollWrapper, ScrollContainer } from 'components/Utils'
-import { Widget } from 'containers/TipicalDailyProfile/DistributionCharts'
+import { ScrollWrapper, ScrollContainer } from '../components/Utils'
+import { Widget } from './TipicalDailyProfile/DistributionCharts'
 
-import { getDailyProfile } from 'services/api'
+import { getDailyProfile } from '../services/api'
 
 import ErrorOutlineIcon from '@material-ui/icons/Error'
