+++ src/containers/TipicalDailyProfile/DistributionByPeriod.jsx
