+++ src/containers/TipicalDailyProfile/DistributionByUserType.jsx
