+++ src/pages/ProductionConsumption.jsx
@@ -6,15 +6,15 @@ import { Grid } from '@material-ui/core'
 import Typography from '@material-ui/core/Typography'
 import Box from '@material-ui/core/Box'
 import Use from '../containers/Generation/Use'
-import KwhBag from 'containers/Generation/KwhBag'
-import Record from 'containers/Generation/Record'
+import KwhBag from '../containers/Generation/KwhBag'
+import Record from '../containers/Generation/Record'
 import { useParams } from 'react-router-dom'
 import { useTranslation } from 'react-i18next'
 import { CsvformatData, kwhRecordToCsvformatData } from '../services/utils'
 import { CSVLink } from 'react-csv'
 import GetAppIcon from '@material-ui/icons/GetApp'
-import GenerationUseContext from 'contexts/GenerationUseContext'
-import ExtraControls from 'components/ExtraControls/ExtraControlsHeader'
+import GenerationUseContext from '../contexts/GenerationUseContext'
+import ExtraControls from '../components/ExtraControls/ExtraControlsHeader'
 import dayjs from 'dayjs'
 
 const useStyles = makeStyles((theme) => ({
